runtime: windows: `get_windows_api_error()` and `windows-err-msg()`
Implement functions to get the last Windows API error. This should be
used when calling Windows API functions to get the proper error.

This is necessary because typically the Windows API does not set
`errno`, and so our typical usage of `strerror(errno)` to get the last
error will yield the ever-so-helpful "No error" error message.