stitcher: emit `.text` before emitting stubs
This ensures we are in a `.text` section before emitting stubs (which
must be in the `.text` section).

Fixes a bug introduced in the previous commit (555eba06) where the
generated `EXPORTS` table would start a `.section drectve` before the
stubs were emitted. Since the stubs did not have a preceding `.text`,
this caused the assembler to interpret the stubs as part of the
`EXPORTS` table, leading to messages like "Warning: .drectve `H(`
unrecognized", and resulting in a malformed executable.