parsed-path: allow `<drive>:/` in Windows paths.
On Windows, paths can be given with either forward or back-slashes (and
sometimes even a mix of the two.) The path parser understands this in
every position, except when checking the drive specifier -- here, it
assumes the drive specifier must be of the form `<drive>:\`. This is
incorrect, `<drive>:/` is also completely acceptable. This causes issues
when these paths are given to Stanza (e.g. via `$STANZA_CONFIG`)

To fix, simply allow `/` as well as `\` when validating the drive
specifier.