build: link statically on Windows
On Windows, link libraries (like `-lpthread`) statically so that the
resulting executable does not require any DLLs at runtime. This is more
pertinent on Windows due to a dependency on `libwinpthread-1.dll` that
results in the compiled executable when linking dynamically. Linking
statically avoids having to ship `libwinpthread-1.dll` and is generally
preferable unless we have a good reason to link dynamically.