Added `defmulti` to list of functions that get annotated in defs-db
This commit causes the definitions database to create `SrcDefMulti`
kinds of definitions. This means that the `defmulti` declarations for a
particular `deftype` will be present in a generated definitions database.

This should help when defining the documentation for a `deftype` interface
as the documentation generator will be able to determine what interface
defmulti's will apply.