Modifies the definition annotation printing to match definition syntax
Previously the definitions that were created by printing the
DefnAnnotation did not have a space between the type parameter
and the argument list. For example, you might get these outputs

```
(a:Int) -> String
<T?>(a:T?) -> T
```

For the first example, there isn't anything inherently wrong.
Bu the second I believe you not compile because there is no space between
the `<>` and the `()`

This adds a prefix space in both cases. This makes it easier to apply
this directly to a function defn signature:

```
defn some-func (a:Int) -> String
defn another<T?> (a:T?) -> T
```